requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.0
colorama==0.4.6
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
